GreenAccounter-backend-monitor-carbon/requirements.txt
pandas
numpy
psutil
argparse
firebase-admin
asyncio
flask
flask-cors
aioelectricitymaps
paramiko
geopandas
python-dotenv